ask/requirements.txt
backcall==0.1.0
decorator==4.2.1
get==1.0.0
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
post==1.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
public==1.0.0
Pygments==2.2.0
query-string==1.0.0
request==1.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
